Decomposed image-preview component
