--- AdobeStockImage/etc/acl.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Acl/etc/acl.xsd">
-    <acl>
-        <resources>
-            <resource id="Magento_Backend::admin">
-                <resource id="Magento_Backend::system">
-                    <resource id="Magento_AdobeStockImage::adobe_stock_image" title="Adobe Stock" translate="title" sortOrder="10">
-                        <resource id="Magento_AdobeStockImage::actions" title="Actions" translate="title">
-                            <resource id="Magento_AdobeStockImage::actions_download" title="Download" translate="title"/>
-                        </resource>
-                    </resource>
-                </resource>
-            </resource>
-        </resources>
-    </acl>
-</config>
