+++ AdobeStockImageAdminUi/view/adminhtml/templates/panel.phtml
@@ -23,12 +23,7 @@
                     "stock_panel": {
                         "component": "Magento_AdobeStockImageAdminUi/panel",
                         "containerId": "#adobe-stock-images-search-modal",
-                        "masonryComponentPath": "adobe_stock_images_listing.adobe_stock_images_listing.adobe_stock_images_columns",
-                        "downloadPreviewUrl": "<?= /* @noEscape */ $block->getUrl('adobe_stock/preview/download') ?>",
-                        "licenseAndDownloadUrl": "<?= /* @noEscape */ $block->getUrl('adobe_stock/license/license') ?>",
-                        "confirmationUrl": "<?= /* @noEscape */ $block->getUrl('adobe_stock/license/confirmation') ?>",
-                        "relatedImagesUrl": "<?= /* @noEscape */ $block->getUrl('adobe_stock/preview/relatedimages') ?>",
-                        "buyCreditsUrl": "<?= /* @noEscape */ $block->getBuyCreditsUrl() ?>"
+                        "masonryComponentPath": "adobe_stock_images_listing.adobe_stock_images_listing.adobe_stock_images_columns"
                     }
                 }
             }
