--- AdobeStockImageAdminUi/view/adminhtml/web/js/config.js
@@ -1,16 +1,0 @@
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-define([], function () {
-    'use strict';
-
-    return {
-        mediaGallerySelector: '.media-gallery-modal:has(#search_adobe_stock)',
-        adobeStockModalSelector: '#adobe-stock-images-search-modal',
-        downloadPreviewUrl: 'adobe_stock/preview/download',
-        licenseAndDownloadUrl: 'adobe_stock/license/license',
-        confirmationUrl: 'adobe_stock/license/confirmation',
-        relatedImagesUrl: 'adobe_stock/preview/relatedimages'
-    };
-});
