+++ AdobeStockImageAdminUi/view/adminhtml/web/template/grid/column/preview/actions.html
@@ -1,0 +1,29 @@
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<ul class="messages" if="getMessages().length > 0">
+    <!-- ko foreach: getMessages() -->
+        <div class="messages">
+            <div attr="class: 'message message-'+code">
+                <div data-ui-id="messages-message-error">
+                    <span data-bind="html: message"></span>
+                </div>
+            </div>
+        </div>
+    <!-- /ko -->
+</ul>
+<button class="action-secondary" type="button" data-bind="visible: !isDownloaded($row()), click: function(){ savePreview() }">
+    <span translate="'Save Preview'"/>
+</button>
+<button class="action-secondary" type="button" data-bind="visible: isDownloaded($row()), click: function(){ locate() }">
+    <span translate="'Locate'"/>
+</button>
+<button class="action-default primary"
+        type="button"
+        attr="title: $t('License and Save')"
+        data-bind="click: function () { licenseProcess(); }">
+    <span translate="'License and Save'"></span>
+</button>
