+++ AdobeStockImageAdminUi/view/adminhtml/web/template/grid/column/preview/related.html
@@ -1,0 +1,64 @@
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<div id="adobe-stock-tabs"
+     class="adobe-stock-tabs"
+     data-bind="mageInit: {
+                    'mage/backend/tabs': {
+                    active: 'series_content_' + $row().id,
+                    destination: '#adobe-stock-tabs-content',
+                    shadowTabs: [],
+                }
+            }">
+    <ul class="tabs-horiz">
+        <li>
+            <a id="series_tab" attr="'href': '#series_content_' + $row().id">
+                <span class="title" translate="'More from this series'"></span>
+            </a>
+        </li>
+        <li>
+            <a id="model_tab" attr="'href': '#model_content_' + $row().id">
+                <span class="title" translate="'More from this model'"></span>
+            </a>
+        </li>
+    </ul>
+</div>
+<div id="adobe-stock-tabs-content" class="adobe-stock-related-images-tab-content">
+    <div attr="'id': 'series_content_' + $row().id">
+        <!-- ko foreach: getSeries($row()) -->
+            <div class="thumbnail" data-bind="click: function(){ $parent.switchImagePreviewToSeriesImage($data, $row()) }">
+                <img attr="src: thumbnail_url, alt: title">
+            </div>
+        <!-- /ko -->
+        <div class="see-more-wrapper" data-bind="click: function(){ seeMoreFromSeries($row()) }">
+            <div class="see-more-content">
+                <div class="three-dots">
+                    <span class="dots"></span>
+                    <span class="dots"></span>
+                    <span class="dots"></span>
+                </div>
+                <a href="" class="see-more">See more</a>
+            </div>
+        </div>
+    </div>
+    <div attr="'id': 'model_content_' + $row().id">
+        <!-- ko foreach: getModel($row()) -->
+            <div class="thumbnail" data-bind="click: function(){ $parent.switchImagePreviewToModelImage($data, $row()) }">
+                <img attr="src: thumbnail_url, alt: title">
+            </div>
+        <!-- /ko -->
+        <div class="see-more-wrapper" data-bind="click: function(){ seeMoreFromModel($row()) }">
+            <div class="see-more-content">
+                <div class="three-dots">
+                    <span class="dots"></span>
+                    <span class="dots"></span>
+                    <span class="dots"></span>
+                </div>
+                <a href="" class="see-more">See more</a>
+            </div>
+        </div>
+    </div>
+</div>
