+++ AdobeUi/view/adminhtml/web/js/components/grid/column/image-preview.js
@@ -5,7 +5,7 @@
 define([
     'underscore',
     'jquery',
-    'Magento_Ui/js/grid/columns/column',
+    'Magento_Ui/js/grid/columns/column'
 ], function (_, $, Column) {
     'use strict';
 
@@ -14,16 +14,48 @@ define([
             previewImageSelector: '[data-image-preview]',
             visibility: [],
             height: 0,
+            previewStyles: {},
+            displayedRecord: {},
             lastOpenedImage: null,
             modules: {
-                masonry: '${ $.parentName }'
+                masonry: '${ $.parentName }',
+                thumbnailComponent: '${ $.parentName }.thumbnail_url',
+            },
+            statefull: {
+                visible: true,
+                sorting: true,
+                lastOpenedImage: true
+            },
+            listens: {
+                '${ $.provider }:params.filters': 'hide',
+                '${ $.provider }:params.search': 'hide'
             }
+        },
+
+        /**
+         * Init observable variables
+         * @return {Object}
+         */
+        initObservable: function () {
+            this._super()
+                .observe([
+                    'visibility',
+                    'height',
+                    'previewStyles',
+                    'displayedRecord',
+                    'lastOpenedImage'
+                ]);
+            this.height.subscribe(function () {
+                this.thumbnailComponent().previewHeight(this.height());
+            }, this);
+
+            return this;
         },
 
         /**
          * Next image preview
          *
-         * @param record
+         * @param {Object} record
          */
         next: function (record) {
             var recordToShow = this.getRecord(record._rowIndex + 1);
@@ -34,7 +66,7 @@ define([
         /**
          * Previous image preview
          *
-         * @param record
+         * @param {Object} record
          */
         prev: function (record) {
             var recordToShow = this.getRecord(record._rowIndex - 1);
@@ -73,6 +105,7 @@ define([
                 img;
 
             this.hide();
+            this.displayedRecord(record);
 
             if (record.rowNumber) {
                 this._selectRow(record.rowNumber);
@@ -88,7 +121,7 @@ define([
             } else {
                 img.load(this._updateHeight.bind(this));
             }
-            this.lastOpenedImage = record;
+            this.lastOpenedImage(record._rowIndex);
         },
 
         /**
@@ -106,7 +139,7 @@ define([
         hide: function () {
             var visibility = this.visibility();
 
-            this.lastOpenedImage = null;
+            this.lastOpenedImage(null);
             visibility.fill(false);
             this.visibility(visibility);
             this.height(0);
@@ -120,8 +153,7 @@ define([
          * @return {*|boolean}
          */
         isVisible: function (record) {
-            if (this.lastOpenedImage
-                && this.lastOpenedImage._rowIndex === record._rowIndex
+            if (this.lastOpenedImage() === record._rowIndex
                 && (
                     this.visibility()[record._rowIndex] === undefined
                     || this.visibility()[record._rowIndex] === false
@@ -130,6 +162,30 @@ define([
                 this.show(record);
             }
             return this.visibility()[record._rowIndex] || false;
+        },
+
+        /**
+         * Get styles for preview
+         *
+         * @returns {Object}
+         */
+        getStyles: function () {
+            this.previewStyles({
+                'margin-top': '-' + this.height()
+            });
+
+            return this.previewStyles();
+        },
+
+        /**
+         * Scroll to preview window
+         */
+        scrollToPreview: function () {
+            $(this.previewImageSelector).get(0).scrollIntoView({
+                behavior: 'smooth',
+                block: 'center',
+                inline: 'nearest'
+            });
         }
     });
 });
