+++ AdobeUi/view/adminhtml/web/js/components/grid/column/overlay.js
@@ -23,6 +23,7 @@ define([
 
                 return styles;
             }
+
             return {};
         }
     });
